Caffeinated edited

+++ store/views.py
@@ -1,4 +1,5 @@
 from .models import Item, OrderItem, Order, Delivery, Collection, Payment
+from account.models import Customer
 from .forms import CaffeinatedForm, PaymentForm, DeliveryForm, CollectionForm
 from django.shortcuts import get_object_or_404, redirect, render
 from django.contrib.auth.decorators import login_required
@@ -49,21 +50,9 @@ def home(request):
     items_main_two = Item.objects.filter(category='Main Course')[:2]
     items_main_four = Item.objects.filter(category='Main Course')[2:4]
     items_sizzlers_two = Item.objects.filter(category='Sizzlers')[:2]    
-    items_sizzlers_four = Item.objects.filter(category='Sizzlers')[2:4]  
+    items_sizzlers_four = Item.objects.filter(category='Sizzlers')[2:4]
 
-    already_in_cart = None
-    customer_order_item = None
-    if request.user.is_authenticated:
-        for item in items_caffeinated_two:
-            order_item = OrderItem.objects.filter(user=request.user, in_cart=True, ordered=False)        
-            if order_item.exists():  
-                already_in_cart = order_item                                                          
-                
-    for item in items_caffeinated_two:            
-        print(item.name, ':', item.customer.filter(username=request.user).exists())
-        
     context = {
-        'already_in_cart': already_in_cart,
         'items_caffeinated_two': items_caffeinated_two, 
         'items_caffeinated_four': items_caffeinated_four, 
         'items_frappe_two': items_frappe_two, 
@@ -91,6 +80,8 @@ def home(request):
 def product_detail(request, item_slug):
     item = Item.objects.get(item_slug=item_slug)                                        
     caffeinated_form = CaffeinatedForm()
+    order_item = OrderItem.objects.filter(user=request.user, item=item, ordered=False, in_cart=True)
+    order_item = order_item.first() 
     
     if request.method == 'POST':
         if request.user.is_authenticated:
@@ -101,6 +92,7 @@ def product_detail(request, item_slug):
                 instance.item = item
                 instance.hot_or_cold = caffeinated_form.cleaned_data['hot_or_cold']           
 
+                # This will set the price whether the customer wants Hot or Cold.
                 if caffeinated_form.cleaned_data['hot_or_cold'] == 'Hot':
                     instance.price = item.hot_price
                 else:
@@ -108,27 +100,41 @@ def product_detail(request, item_slug):
 
                 instance.hot_cold = caffeinated_form.cleaned_data['hot_or_cold']           
 
+                # This will check the checkbox if the customer would like to add an add-ons.
+                # Also to add the price and total it.
                 if caffeinated_form.cleaned_data['milk']:
+                    instance.total_price += item.caffeinated_add.milk
                     instance.milk = True
-                elif caffeinated_form.cleaned_data['whip_cream']:
+                if caffeinated_form.cleaned_data['whip_cream']:
+                    instance.total_price += item.caffeinated_add.whip_cream
                     instance.whip_cream = True
-                elif caffeinated_form.cleaned_data['syrup_pump']:
+                if caffeinated_form.cleaned_data['syrup_pump']:
+                    instance.total_price += item.caffeinated_add.syrup_pump
                     instance.syrup_pump = True
-                elif caffeinated_form.cleaned_data['espresso_shot']:
+                if caffeinated_form.cleaned_data['espresso_shot']:
                     instance.espresso_shot = True     
+                    instance.total_price += item.caffeinated_add.espresso_shot
 
-                add_on_list = [instance.milk, instance.whip_cream, instance.syrup_pump, instance.espresso_shot]
+                if order_item:
+                    # If the customer go back again on the specific product and he or she click again the "Add to Cart" button.
+                    # Then it will check first if he or she still wants to buy the same add-ons.
+                    # If he or she is, then it will just increment the quantity of this product.
+                    if instance.hot_or_cold == order_item.hot_or_cold and order_item.milk == instance.milk and order_item.whip_cream == instance.whip_cream and order_item.syrup_pump == instance.syrup_pump and order_item.espresso_shot == instance.espresso_shot:
+                        order_item.quantity += 1
+                        order_item.save()
+                        return redirect('store:cart')  
+                    # But if not, then it will add a new and the same product with the different add-ons that he or she choose.
+                    else:  
+                        instance.total_price += instance.price
+                        order, created = Order.objects.get_or_create(user=request.user, ordered=False)                 
+                        order.items.add(caffeinated_form.save())   
+                        instance.in_cart = True   
+                        item.customer.add(request.user)
+                        item.save()                                              
+                        instance.save()                                       
+                        return redirect('store:cart')                                                    
 
-                for data in range(4):
-                    if data == 0 and add_on_list[0]:
-                        instance.total_price += item.caffeinated_add.milk
-                    elif data == 1 and add_on_list[1]:
-                        instance.total_price += item.caffeinated_add.whip_cream
-                    elif data == 2 and add_on_list[2]:
-                        instance.total_price += item.caffeinated_add.syrup_pump
-                    elif data == 3 and add_on_list[3]:
-                        instance.total_price += item.caffeinated_add.espresso_shot
-
+                # This will add to OrderItem table and input whatever the customer added.
                 instance.total_price += instance.price
                 order, created = Order.objects.get_or_create(user=request.user, ordered=False)                 
                 order.items.add(caffeinated_form.save())   
@@ -136,9 +142,6 @@ def product_detail(request, item_slug):
                 item.customer.add(request.user)
                 item.save()                                              
                 instance.save()                                       
-                print('Success')
-            else:
-                print('Error')
 
             return redirect('store:cart')     
         else:
@@ -164,13 +167,14 @@ def product_detail(request, item_slug):
     context = {'item': item, 'caffeinated_form': caffeinated_form, 'order_quantity': order_quantity(request)}
     return render(request, 'store/product-detail.html', context)
 
-def product_update(request, item_slug):
+def product_update(request, item_slug, order_item_id):
+    print('ID:', order_item_id)
     item = Item.objects.get(item_slug=item_slug)
-    order_item = OrderItem.objects.get(user=request.user, item=item, ordered=False)
+    order_item = OrderItem.objects.get(id=order_item_id, user=request.user, item=item, ordered=False)
+    
     caffeinated_form = CaffeinatedForm(instance=order_item)
     if request.method == 'POST':
         caffeinated_form = CaffeinatedForm(request.POST, instance=order_item)
-        order_qs = Order.objects.filter(user=request.user, ordered=False)            
         if caffeinated_form.is_valid():
             instance = caffeinated_form.save(commit=False)
             instance.hot_or_cold = caffeinated_form.cleaned_data['hot_or_cold']           
@@ -181,7 +185,8 @@ def product_update(request, item_slug):
                 instance.price = item.cold_price
 
             instance.hot_cold = caffeinated_form.cleaned_data['hot_or_cold']           
-            
+
+            # This update among add-ons.
             if caffeinated_form.cleaned_data['milk']:
                 instance.milk = True
             elif caffeinated_form.cleaned_data['whip_cream']:
@@ -201,8 +206,8 @@ def product_update(request, item_slug):
                 elif data == 2 and add_on_list[2]:
                     instance.total_price += item.caffeinated_add.syrup_pump
                 elif data == 3 and add_on_list[3]:
-                    instance.total_price += item.caffeinated_add.espresso_shot
-                    
+                    instance.total_price += item.caffeinated_add.espresso_shot            
+                
             instance.total_price += instance.price                    
             instance.save()                                                    
         else:
@@ -210,7 +215,7 @@ def product_update(request, item_slug):
             
         return redirect('store:cart')
     
-    context = {'item':item, 'caffeinated_form': caffeinated_form, 'order_quantity': order_quantity(request)}
+    context = {'item':item, 'order_item': order_item, 'caffeinated_form': caffeinated_form, 'order_quantity': order_quantity(request)}
     return render(request, 'store/product-update.html', context)
 
 def all_product(request):
@@ -279,7 +284,7 @@ def cart(request):
     
     if order.exists():
         customer_order = order.first()
-        all_order = customer_order.items.all().order_by('-id')              
+        all_order = customer_order.items.all().order_by('-id')  
         # Delete the customer in Order table if he or she has no order left in cart.
         order_count = customer_order.items.all().count()                      
         if order_count == 0:
@@ -291,14 +296,12 @@ def cart(request):
         
     return render(request, 'store/cart.html', context)
 
-def update_item(request, item_slug):
-    pass
-
 @login_required
-def decrease_quantity(request, item_slug): 
+def decrease_quantity(request, item_slug, order_item_id): 
     item = get_object_or_404(Item, item_slug=item_slug)
-    order_item, created = OrderItem.objects.get_or_create(user=request.user, item=item, ordered=False)
+    order_item = OrderItem.objects.get(user=request.user, id=order_item_id, item=item, ordered=False)
     order_qs = Order.objects.filter(user=request.user, ordered=False)
+    order_item_count = OrderItem.objects.filter(user=request.user, item=item, ordered=False).count()
     
     if order_qs.exists():
         order = order_qs[0]    
@@ -306,17 +309,21 @@ def decrease_quantity(request, item_slug):
         if order.items.filter(item__item_slug=item.item_slug).exists():
             order_item.quantity -= 1
             order_item.save()
-            # If the order item quantity is zero then delete in order item table.
-            if order_item.quantity == 0:
+            # When the order item quantity is zero then delete in order item table.
+            # If the customer has only 1 the same product left then delete the Item object in OrderItem Table and remove also the customer to that product. 
+            if order_item.quantity == 0 and order_item_count == 1:
                 order_item.delete()     
                 item.customer.remove(request.user)                   
+            # If not, then remove only the Item in OrderItem table.
+            if order_item.quantity == 0 and order_item_count > 1:
+                order_item.delete()                     
                 
     return redirect('store:cart')         
 
 @login_required
-def increase_quantity(request, item_slug): 
+def increase_quantity(request, item_slug, order_item_id): 
     item = get_object_or_404(Item, item_slug=item_slug)
-    order_item, created = OrderItem.objects.get_or_create(user=request.user, item=item, ordered=False)
+    order_item, created = OrderItem.objects.get_or_create(user=request.user, id=order_item_id, item=item, ordered=False)
     order_qs = Order.objects.filter(user=request.user, ordered=False)
     if order_qs.exists():
         order = order_qs[0]    
@@ -333,11 +340,21 @@ def increase_quantity(request, item_slug):
     return redirect('store:cart')   
      
 @login_required
-def remove_product(request, item_slug):
-    item = get_object_or_404(Item, item_slug=item_slug)    
-    item.customer.remove(request.user)                       
-    OrderItem.objects.get(user=request.user, item=item, ordered=False).delete()
-    
+def remove_product(request, item_slug, order_item_id):
+    item = get_object_or_404(Item, item_slug=item_slug)        
+    order_item = OrderItem.objects.get(user=request.user, id=order_item_id, item=item, ordered=False)    
+    order_item_count = OrderItem.objects.filter(user=request.user, item=item, ordered=False).count()    
+    print('order_item quantity:', order_item.quantity)
+    print('order_item_count:', order_item_count)
+    # When the order item quantity is zero then delete in order item table.
+    # If the customer has only 1 the same product left then delete the Item object in OrderItem Table and remove also the customer to that product. 
+    if order_item_count == 1:
+        item.customer.remove(request.user)                       
+        OrderItem.objects.get(user=request.user, id=order_item_id, item=item, ordered=False).delete()
+    # If not, then remove only the Item in OrderItem table.        
+    if order_item_count > 1:
+        order_item.delete()
+        
     return redirect('store:cart')       
 
 @login_required
@@ -346,7 +363,7 @@ def delivery(request):
     delivery_fee = payment.first().delivery_fee
     delivery_form = DeliveryForm()
     order = Order.objects.filter(user=request.user, ordered=False) 
-    order_items = OrderItem.objects.filter(user=request.user, ordered=False) 
+    order_items = OrderItem.objects.filter(user=request.user, ordered=False, in_cart=True) 
 
     # If the customer has no order yet.
     if not order.exists():
@@ -356,7 +373,6 @@ def delivery(request):
     # Get the customer orders. Calculate subtotal and total.
     customer_order = order.first() 
     customer_items = order_items.all() 
-    print('Customer order:', customer_items)
     
     all_order = customer_order.items.all().order_by('-id') 
     total = delivery_fee + sub_total(request)
@@ -365,7 +381,6 @@ def delivery(request):
     if request.method == 'POST':
         delivery_form = DeliveryForm(request.POST)       
         office_home = request.POST.get('home-office') 
-        print('Office home:', office_home)                         
         if delivery_form.is_valid():
             instance = delivery_form.save(commit=False)
             instance.user = request.user            
@@ -376,7 +391,11 @@ def delivery(request):
                 # This will make the ordered field equals to True in OrderItem table.
                 for product in customer_items:
                     product.ordered = True
+                    product.in_cart = False
                     product.save()
+                # Remove all the product that customer ordered after placing order.
+                for product in Item.objects.filter(customer=request.user):
+                    product.customer.remove(request.user)
                 customer_order.save()
                 instance.save()
                 Payment.objects.create(user=request.user, delivery=delivery_form.instance, order=customer_order, subtotal=subtotal, delivery_fee=delivery_fee, total=total)  
@@ -396,9 +415,7 @@ def collection(request):
     # Get the customer orders. Calculate subtotal and total.    
     # The subtotal and total is the same due to had no delivery fee charge.
     customer_order = order.first()    
-    customer_items = order_items.all() 
-    for data in customer_items:
-        print('Data:', data)     
+    customer_items = order_items.all()   
     all_order = customer_order.items.all().order_by('-id')    
     total = sub_total(request)
